requirements.txt
streamlit
numpy
scikit-learn
Pillow